+++ lib/src/show_custom_loader.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_loading_kit/src/set_custom_loader.dart';
 
 import 'custom_loader_widget.dart';
@@ -7,23 +7,23 @@ import 'custom_loader_widget.dart';
 class AnimatedLoader{
 
   final BuildContext context;
-  final height;
-  final width;
-  bool isDismissable = true;
-  bool isPoped = false;
+  final double height;
+  final double width;
+  bool isDismissible;
+  bool isPopped = false;
 
-  static final String toasterDialog = 'assets/toster.json';
-  static final String basketBall = 'assets/basket_ball.json';
-  static final String plane = 'assets/plane.json';
-  static final String liquidBubble = 'assets/liquid_bubble.json';
-  static final String loadingRobot = 'assets/loading_robot.json';
-  static final String robotHello = 'assets/robot_hello.json';
-  static final String paymentCard = 'assets/payment_card.json';
-  static final String chargingIndicator = 'assets/charging_indicator.json';
-  static final String launchRocket = 'assets/launch_rocket.json';
-  static final String orangeEmoji = 'assets/orange_emoji.json';
+  static const String TOASTER_DIALOG = 'assets/toster.json';
+  static const String BASKETBALL = 'assets/basket_ball.json';
+  static const String PLANE = 'assets/plane.json';
+  static const String LIQUID_BUBBLE = 'assets/liquid_bubble.json';
+  static const String ROBOT_WAIT = 'assets/loading_robot.json';
+  static const String ROBOT_HELLO = 'assets/robot_hello.json';
+  static const String PAYMENT_CARD = 'assets/payment_card.json';
+  static const String CHARGER = 'assets/charging_indicator.json';
+  // static const String LAUNCH_ROCKET = 'assets/launch_rocket.json';
+  // static const String ORANGE_EMOJI = 'assets/orange_emoji.json';
 
-  AnimatedLoader({this.isDismissable,this.context,this.height,this.width});
+  AnimatedLoader({ this.isDismissible = true, required this.context, this.height = 100.0, this.width = 100.0});
 
   void showDialog(String assets){
     SetCustomLoader customDialog = new SetCustomLoader(
@@ -32,7 +32,7 @@ class AnimatedLoader{
           assets:assets,
           height: height,
           width: width,
-          onTap: isDismissable == null || isDismissable ? (){isPoped = true;Navigator.of(context).pop();}:(){},
+          onTap: isDismissible ? (){isPopped = true;Navigator.of(context).pop();}:(){},
         )
     );
     customDialog.setDialog();
@@ -40,7 +40,7 @@ class AnimatedLoader{
 
   void removeDialog(){
 
-    if(isPoped){
+    if(isPopped){
     }else{
       Navigator.of(context).pop();
     }