+++ CHANGELOG.md
@@ -1,3 +1,13 @@
-## [0.0.1] - Release
+## [0.0.1] - 22/05/2021
+
+Flutter Loading Package Initial Release
+
+## [0.0.2] - 22/05/2021
 
 Flutter Loading Package
+Minor Updates
+
+## [0.0.3] - 22/05/2021
+
+- Added Null Safety
+- Removed few Animated Loaders